--- src/app/(providers)/(root)/[workspaceId]/channels/(chat)/[id]/(home)/layout.tsx
@@ -1,7 +1,0 @@
-import { StrictPropsWithChildren } from '@/types/common';
-
-const ChatDetailHomeLayout = ({ children }: StrictPropsWithChildren) => {
-  return <>{children}</>;
-};
-
-export default ChatDetailHomeLayout;
