+++ src/app/(providers)/(root)/[workspaceId]/channels/add/(home)/page.tsx
@@ -2,7 +2,6 @@
 
 import useWorkspaceId from '@/hooks/useWorkspaceId';
 import { useSearchParams } from 'next/navigation';
-import { useSearchUsers } from '../_provider/SearchUsersProvider';
 import useCreateChannel from '../_hooks/useCreateChannel';
 import { useRouter } from 'next/navigation';
 import { isEmpty } from '@/utils/isEmpty';
@@ -10,20 +9,23 @@ import { CHANNEL_TYPE } from '@/constants/channel';
 import { fetchExistingChannelId } from '../_utils/fetchExistingChannelId';
 import AddChannelLayout from '../_components/AddChannelLayout';
 import { SearchInput, SearchResults, SelectedUsers } from '../_components';
+import useChatSearchUsersStore from '@/store/chatSearchUserStore';
+import { useWorkspaceUserId } from '@/hooks/useWorkspaceUserId';
 
 const AddChatPage = () => {
   const workspaceId = useWorkspaceId();
   const searchParams = useSearchParams();
+  const workspaceUserId = useWorkspaceUserId();
+  const router = useRouter();
 
   const type = searchParams.get('type');
-  const { getSelectedUserIds } = useSearchUsers();
+  const { getSelectedUserIds } = useChatSearchUsersStore();
   const { handleCreateChannelAndUsers } = useCreateChannel();
-  const router = useRouter();
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
-    const userIds = getSelectedUserIds();
+    const userIds = getSelectedUserIds(workspaceUserId);
     if (isEmpty(userIds)) return;
 
     const isGroupChat = userIds.length > 2 || type === CHANNEL_TYPE.video;
