--- src/app/(providers)/(root)/[workspaceId]/channels/add/layout.tsx
@@ -1,8 +1,0 @@
-import type { StrictPropsWithChildren } from '@/types/common';
-import { SearchUsersProvider } from './_provider/SearchUsersProvider';
-
-const AddChatProviderLayout = ({ children }: StrictPropsWithChildren) => {
-  return <SearchUsersProvider>{children}</SearchUsersProvider>;
-};
-
-export default AddChatProviderLayout;
